Se reordenan responsabilidades de los modulos data_loading y exploracion. El modulo rentabilidad genera reportes siguiendo la estructura modular. Rutas asignadas con exito
