+++ src/exportacion.py
@@ -1,0 +1,2 @@
+def exportar_excel(df, ruta_salida):
+    df.to_excel(ruta_salida, index=False)
